requirements.txt
Django==1.9.13
djangorestframework==3.7.1